modity integer to float

+++ temperature.py
@@ -1,3 +1,3 @@
 celsius = input('請輸入攝式溫度：')
-fahrenheit = int(celsius) * (9/5) + 32
+fahrenheit = float(celsius) * (9/5) + 32
 print('華式溫度：', fahrenheit)
